Create stopthis
If we are to consider every day our last, I would want to let people know about the things in this list. The Oil and Gas industry is destructive, insidious and must be prevented from getting away with the things it is doing in broad daylight. The era when governments should be working with the industry should be over.